requirements.txt
cement==3.0.2
jinja2
pyyaml
colorlog
boto3
psycopg2-binary
pymssql

